Ui auth updates

+++ src/features/ticket/actions/delete-ticket.ts
@@ -4,12 +4,30 @@ import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 
 import { setCookieByKey } from '@/actions/cookies'
-import { fromErrorToActionState } from '@/components/form/utils/to-action-state'
+import {
+  fromErrorToActionState,
+  toActionState,
+} from '@/components/form/utils/to-action-state'
+import { getAuthOrRedirect } from '@/features/auth/queries/get-auth-or-redirect'
+import { isOwner } from '@/features/auth/utils/is-owner'
 import { prisma } from '@/lib/prisma'
 import { ticketsPath } from '@/paths'
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect()
+
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: { id: id },
+    })
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState(
+        'ERROR',
+        'You are not authorized to edit this ticket'
+      )
+    }
+
     await prisma.ticket.delete({
       where: { id: id },
     })
